fixed usage of tokenlists refactored alchemy history fetching
