+++ src/lib/hooks/use-token-history.ts
@@ -1,12 +1,11 @@
 import { useQuery } from '@tanstack/react-query'
-import { AssetTransfersWithMetadataResult } from 'alchemy-sdk'
-import { Address, zeroAddress } from 'viem'
+import { Address } from 'viem'
 import { usePublicClient } from 'wagmi'
 
 import { getLeverageAction } from '@/app/leverage/utils/get-leverage-type'
 import { useNetwork } from '@/lib/hooks/use-network'
 import { useWallet } from '@/lib/hooks/use-wallet'
-import { isContract } from '@/lib/utils'
+import { AssetTransfersWithMetadataResult } from 'alchemy-sdk'
 
 export const useTokenHistory = (...tokens: (string | Address)[]) => {
   const { address: user } = useWallet()
@@ -32,33 +31,11 @@ export const useTokenHistory = (...tokens: (string | Address)[]) => {
       })
 
       const transfers =
-        (await response.json()) as AssetTransfersWithMetadataResult[]
+        (await response.json()) as (AssetTransfersWithMetadataResult & {
+          action: ReturnType<typeof getLeverageAction>
+        })[]
 
-      return Promise.all(
-        transfers.map(async (transfer) => {
-          const isMint = transfer.from === zeroAddress
-          const isBurn = transfer.to === zeroAddress
-          const isFromUser = transfer.from === user?.toLowerCase()
-          const isToUser = transfer.to === user?.toLowerCase()
-
-          const isFromContract = await isContract(client!, transfer.from)
-          const isToContract = await isContract(client!, transfer.to!)
-
-          const action = getLeverageAction({
-            isMint,
-            isBurn,
-            isFromContract,
-            isToContract,
-            isFromUser,
-            isToUser,
-          })
-
-          return {
-            ...transfer,
-            action,
-          }
-        }),
-      )
+      return transfers
     },
     select: (data) =>
       data.sort(
